(#1074) Rename UpperText
