+++ src/test/java/org/cactoos/map/SolidTest.java
@@ -28,7 +28,7 @@ import org.cactoos.Scalar;
 import org.cactoos.iterable.IterableOf;
 import org.cactoos.text.SubText;
 import org.cactoos.text.TextOf;
-import org.cactoos.text.UpperText;
+import org.cactoos.text.Upper;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.collection.IsMapContaining;
 import org.hamcrest.core.AllOf;
@@ -131,7 +131,7 @@ public final class SolidTest {
     public void mapsIterableWithKeyFuncAndValueFunc() {
         final Solid<String, String> map = new Solid<>(
             key -> new SubText(new TextOf(key), 0, 1).asString(),
-            value -> new UpperText(new TextOf(value)).asString(),
+            value -> new Upper(new TextOf(value)).asString(),
             new IterableOf<>("aa", "bb")
         );
         MatcherAssert.assertThat(
@@ -157,7 +157,7 @@ public final class SolidTest {
     public void mapsEmptyIterableWithKeyFuncAndValueFunc() {
         final Solid<String, String> map = new Solid<>(
             key -> new SubText(new TextOf(key), 0, 1).asString(),
-            value -> new UpperText(new TextOf(value)).asString(),
+            value -> new Upper(new TextOf(value)).asString(),
             new IterableOf<String>()
         );
         MatcherAssert.assertThat(
@@ -175,7 +175,7 @@ public final class SolidTest {
             new Solid<>(
                 entry -> new MapEntry<>(
                     new SubText(new TextOf(entry), 0, 1).asString(),
-                    new UpperText(new TextOf(entry)).asString()
+                    new Upper(new TextOf(entry)).asString()
                 ),
                 new IterableOf<>("aa", "bb")
             ),
@@ -202,7 +202,7 @@ public final class SolidTest {
             new Solid<>(
                 entry -> new MapEntry<>(
                     new SubText(new TextOf(entry), 0, 1).asString(),
-                    new UpperText(new TextOf(entry)).asString()
+                    new Upper(new TextOf(entry)).asString()
                 ),
                 new IterableOf<String>()
             ),
